ml/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "clean-commits"
version = "1.0.0"
requires-python = ">=3.12, <3.14"
authors = [{ name = "Andrew", email = "hello.andrew.yan@outlook.com" }]

dependencies = [
  "torch",
  "pandas==3.0.1",
  "sentence-transformers==5.3.0",
  "faiss-cpu==1.13.2",
  "rich==14.3.3",
  "python-dotenv==1.2.2",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" },
]